Bug fix: Invalid System.Runtime reference in Number.il
ILAsm could not resolve the System.Runtime assembly properly in Numbers.il.
Swapping the order of the main assmebly .il (disassembled) file and other .il files in the arguments list to ILAsm seems to fixed the issue.

That's likely due to ILAsm is processing the arguments in the received order and thus not yet loading the correct System.Runtime.